libpod: make removePodCgroup linux specific
